--- a/Gruntfile.js
+++ b/Gruntfile.js
@@ -1,14 +1,8 @@
 module.exports = function (grunt) {
     // Load the plugin that provides the tasks
-    grunt.loadNpmTasks('grunt-contrib-jshint');
-    grunt.loadNpmTasks('grunt-contrib-concat');
-    grunt.loadNpmTasks('grunt-contrib-uglify');
-    grunt.loadNpmTasks('grunt-contrib-cssmin');
-    grunt.loadNpmTasks('grunt-contrib-clean');
-    grunt.loadNpmTasks('grunt-contrib-watch');
-    grunt.loadNpmTasks('grunt-contrib-copy');
-    grunt.loadNpmTasks('grunt-contrib-sass');
-    grunt.loadNpmTasks('grunt-spritesmith');
+    require('jit-grunt')(grunt, {
+        sprite: 'grunt-spritesmith'              
+    });
 
     // Project configuration.
     grunt.initConfig({