--- a/package.json
+++ b/package.json
@@ -41,7 +41,8 @@
     "grunt-contrib-uglify": "^0.9.2",
     "grunt-contrib-watch": "^0.6.1",
     "grunt-spritesmith": "^5.0.0",
-    "nodemon": "^1.5.0"
+    "nodemon": "^1.5.0",
+    "jit-grunt": "~0.9.1"
   },
   "dependencies": {
     "express": "^4.13.3",