A few changes to improve performance when using a vagrant VM :
- use of nfs synced folders on mac and linux : grunt watch move from 40 seconds to about 15 seconds
- use of smb synced folders on windows : needs testing, i don't have a working windows machine available
- use of it-grunt for plugin loading : grunt watch change from 15 seconds to about 5 seconds

TODO : test & check improvement on windows with smb synced folders