@Anicky 2015-10-03
On windows, Vagrant doesn't start with 'smb' type.