@lucaslegoff 2015-10-03
Well, the documentation says differently : 
>"SMB is built-in to Windows machines and provides a higher performance alternative to some other mechanisms such as VirtualBox shared folders."

https://docs.vagrantup.com/v2/synced-folders/smb.html